+++ packages/m2m-gateway/test/api/tenants/assignTenantCertifiedAttribute.test.ts
@@ -8,6 +8,7 @@ import {
 import { generateId, pollingMaxRetriesExceeded } from "pagopa-interop-models";
 import { AuthRole, authRole } from "pagopa-interop-commons";
 import request from "supertest";
+import { m2mGatewayApi } from "pagopa-interop-api-clients";
 import { api, mockTenantService } from "../../vitest.api.setup.js";
 import { appBasePath } from "../../../src/config/appBasePath.js";
 import {
@@ -16,27 +17,29 @@ import {
 } from "../../../src/model/errors.js";
 import { toM2MGatewayApiTenantCertifiedAttribute } from "../../../src/api/tenantApiConverter.js";
 
-describe("DELETE /tenants/:tenantId/certifiedAttributes/:attributeId router test", () => {
+describe("POST /tenants/:tenantId/certifiedAttributes router test", () => {
   const mockApiResponse = getMockedApiCertifiedTenantAttribute();
   const mockResponse = toM2MGatewayApiTenantCertifiedAttribute(mockApiResponse);
 
-  const makeRequest = async (token: string) =>
+  const makeRequest = async (
+    token: string,
+    body: m2mGatewayApi.TenantCertifiedAttributeSeed
+  ) =>
     request(api)
-      .delete(
-        `${appBasePath}/tenants/${generateId()}/certifiedAttributes/${generateId()}`
-      )
-      .set("Authorization", `Bearer ${token}`);
+      .post(`${appBasePath}/tenants/${generateId()}/certifiedAttributes`)
+      .set("Authorization", `Bearer ${token}`)
+      .send(body);
 
   const authorizedRoles: AuthRole[] = [authRole.M2M_ADMIN_ROLE];
   it.each(authorizedRoles)(
     "Should return 200 and perform service calls for user with role %s",
     async (role) => {
-      mockTenantService.revokeCertifiedAttribute = vi
+      mockTenantService.assignTenantCertifiedAttribute = vi
         .fn()
         .mockResolvedValue(mockResponse);
 
       const token = generateToken(role);
-      const res = await makeRequest(token);
+      const res = await makeRequest(token, { id: generateId() });
 
       expect(res.status).toBe(200);
       expect(res.body).toEqual(mockResponse);
@@ -47,8 +50,17 @@ describe("DELETE /tenants/:tenantId/certifiedAttributes/:attributeId router test
     Object.values(authRole).filter((role) => !authorizedRoles.includes(role))
   )("Should return 403 for user with role %s", async (role) => {
     const token = generateToken(role);
-    const res = await makeRequest(token);
+    const res = await makeRequest(token, { id: generateId() });
     expect(res.status).toBe(403);
+  });
+
+  it("Should return 400 if passed an invalid seed", async () => {
+    const token = generateToken(authRole.M2M_ADMIN_ROLE);
+    const res = await makeRequest(token, {
+      invalidParam: "invalidValue",
+    } as unknown as m2mGatewayApi.TenantCertifiedAttributeSeed);
+
+    expect(res.status).toBe(400);
   });
 
   it.each([
@@ -56,11 +68,11 @@ describe("DELETE /tenants/:tenantId/certifiedAttributes/:attributeId router test
     missingMetadata(),
     pollingMaxRetriesExceeded(3, 10),
   ])("Should return 500 in case of $code error", async (error) => {
-    mockTenantService.revokeCertifiedAttribute = vi
+    mockTenantService.assignTenantCertifiedAttribute = vi
       .fn()
       .mockRejectedValue(error);
     const token = generateToken(authRole.M2M_ADMIN_ROLE);
-    const res = await makeRequest(token);
+    const res = await makeRequest(token, { id: generateId() });
 
     expect(res.status).toBe(500);
   });
@@ -73,11 +85,11 @@ describe("DELETE /tenants/:tenantId/certifiedAttributes/:attributeId router test
   ])(
     "Should return 500 when API model parsing fails for response",
     async (resp) => {
-      mockTenantService.revokeCertifiedAttribute = vi
+      mockTenantService.assignTenantCertifiedAttribute = vi
         .fn()
         .mockResolvedValue(resp);
       const token = generateToken(authRole.M2M_ADMIN_ROLE);
-      const res = await makeRequest(token);
+      const res = await makeRequest(token, { id: generateId() });
 
       expect(res.status).toBe(500);
     }
