+++ packages/m2m-gateway/test/api/tenants/revokeTenantCertifiedAttribute.test.ts
@@ -8,7 +8,6 @@ import {
 import { generateId, pollingMaxRetriesExceeded } from "pagopa-interop-models";
 import { AuthRole, authRole } from "pagopa-interop-commons";
 import request from "supertest";
-import { m2mGatewayApi } from "pagopa-interop-api-clients";
 import { api, mockTenantService } from "../../vitest.api.setup.js";
 import { appBasePath } from "../../../src/config/appBasePath.js";
 import {
@@ -17,29 +16,27 @@ import {
 } from "../../../src/model/errors.js";
 import { toM2MGatewayApiTenantCertifiedAttribute } from "../../../src/api/tenantApiConverter.js";
 
-describe("POST /tenants/:tenantId/certifiedAttributes router test", () => {
+describe("DELETE /tenants/:tenantId/certifiedAttributes/:attributeId router test", () => {
   const mockApiResponse = getMockedApiCertifiedTenantAttribute();
   const mockResponse = toM2MGatewayApiTenantCertifiedAttribute(mockApiResponse);
 
-  const makeRequest = async (
-    token: string,
-    body: m2mGatewayApi.TenantCertifiedAttributeSeed
-  ) =>
+  const makeRequest = async (token: string) =>
     request(api)
-      .post(`${appBasePath}/tenants/${generateId()}/certifiedAttributes`)
-      .set("Authorization", `Bearer ${token}`)
-      .send(body);
+      .delete(
+        `${appBasePath}/tenants/${generateId()}/certifiedAttributes/${generateId()}`
+      )
+      .set("Authorization", `Bearer ${token}`);
 
   const authorizedRoles: AuthRole[] = [authRole.M2M_ADMIN_ROLE];
   it.each(authorizedRoles)(
     "Should return 200 and perform service calls for user with role %s",
     async (role) => {
-      mockTenantService.assignCertifiedAttribute = vi
+      mockTenantService.revokeTenantCertifiedAttribute = vi
         .fn()
         .mockResolvedValue(mockResponse);
 
       const token = generateToken(role);
-      const res = await makeRequest(token, { id: generateId() });
+      const res = await makeRequest(token);
 
       expect(res.status).toBe(200);
       expect(res.body).toEqual(mockResponse);
@@ -50,17 +47,8 @@ describe("POST /tenants/:tenantId/certifiedAttributes router test", () => {
     Object.values(authRole).filter((role) => !authorizedRoles.includes(role))
   )("Should return 403 for user with role %s", async (role) => {
     const token = generateToken(role);
-    const res = await makeRequest(token, { id: generateId() });
+    const res = await makeRequest(token);
     expect(res.status).toBe(403);
-  });
-
-  it("Should return 400 if passed an invalid seed", async () => {
-    const token = generateToken(authRole.M2M_ADMIN_ROLE);
-    const res = await makeRequest(token, {
-      invalidParam: "invalidValue",
-    } as unknown as m2mGatewayApi.TenantCertifiedAttributeSeed);
-
-    expect(res.status).toBe(400);
   });
 
   it.each([
@@ -68,11 +56,11 @@ describe("POST /tenants/:tenantId/certifiedAttributes router test", () => {
     missingMetadata(),
     pollingMaxRetriesExceeded(3, 10),
   ])("Should return 500 in case of $code error", async (error) => {
-    mockTenantService.assignCertifiedAttribute = vi
+    mockTenantService.revokeTenantCertifiedAttribute = vi
       .fn()
       .mockRejectedValue(error);
     const token = generateToken(authRole.M2M_ADMIN_ROLE);
-    const res = await makeRequest(token, { id: generateId() });
+    const res = await makeRequest(token);
 
     expect(res.status).toBe(500);
   });
@@ -85,11 +73,11 @@ describe("POST /tenants/:tenantId/certifiedAttributes router test", () => {
   ])(
     "Should return 500 when API model parsing fails for response",
     async (resp) => {
-      mockTenantService.assignCertifiedAttribute = vi
+      mockTenantService.revokeTenantCertifiedAttribute = vi
         .fn()
         .mockResolvedValue(resp);
       const token = generateToken(authRole.M2M_ADMIN_ROLE);
-      const res = await makeRequest(token, { id: generateId() });
+      const res = await makeRequest(token);
 
       expect(res.status).toBe(500);
     }
